dead end removal
Added a flag for networks that are intended to be used as an output of the whole circuit. It's currently a little bothersome to be accessed though:
networks[networkLoukUp[somewire.source.source]].markAsOutput();

Compilation goes through all of the circuit and only outputs circuits that could possibly influence the wires marked as output. It's thus sufficient to only mark the very last thing as such.