Complete core rewrite and simulator addition
Rewritten combiler from the ground up in order to make it more well-structured - this includes separating it into a header and an implementation part.
Also used the thus gained clarity to hack in support for simulation of the whole circuit - simply rerun the circuit generating code again, and it'll instead simulate the circuit.